Added Dashboard. Moved stocksearch from front-end component to redux slice
Worked on feedback from Hodlnaut final interview. Simon pointed out that my codes were not very "Redux-y" because the Stock API call was in the frontend.

Moved it to the redux slice. Turns out the issue was in passing information down (the second params in the CreateAsyncThunk is NOT for another params, it's specifically for APIThunk)